requirements.txt
Django>=4.0.6,<4.1
djangorestframework>=3.13.1,<3.14
pytest-django>=4.5.0,<4.6
gunicorn>=20.1.0,<20.2